+++ internal/global/status_snapshot.go
@@ -55,8 +55,7 @@ func BuildStatusSnapshot(engine string) ([]byte, error) {
 				"k8s":      BoolState(CheckContainer(engine, "kind-control-plane")),
 				"jwt":      BoolState(CheckContainer(engine, "hal-gitlab")),
 				"ldap":     BoolState(CheckContainer(engine, "hal-openldap")),
-				"database": BoolState(CheckContainer(engine, "hal-vault-mariadb")),
-				"oracle":   BoolState(CheckContainer(engine, "hal-vault-oracle-db")),
+				"database": BoolState(CheckContainer(engine, "hal-vault-mariadb") || CheckContainer(engine, "hal-vault-oracle-db")),
 				"oidc":     BoolState(CheckContainer(engine, "hal-keycloak")),
 			},
 			"http://vault.localhost:8200"),
